# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Jatek Food Delivery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ads:
    get:
      operationId: listAds
      tags:
        - content
      summary: List active ads/promos
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"
...
  /backend/categories/all:
    get:
      operationId: listBackendCategoriesAll
      tags:
        - backend
      summary: List all categories (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryWithSubs"
...
  /backend/shorts:
    get:
      operationId: listBackendShorts
      tags:
        - backend
      summary: List all shorts (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: shorts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Short"
    post:
      operationId: createBackendShort
      tags:
        - backend
      summary: Create a short (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShortBody"
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Short"
  /backend/shorts/{id}:
    patch:
      operationId: updateBackendShort
      tags:
        - backend
      summary: Update a short (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShortBody"
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Short"
    delete:
      operationId: deleteBackendShort
      tags:
        - backend
      summary: Delete a short (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: deleted
components:
  schemas:
...
    Category:
      type: object
      required:
        - id
        - name
        - slug
        - icon
        - accentColor
        - businessType
        - isActive
        - sortOrder
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        accentColor:
          type: string
        parentId:
          type:
            - integer
            - "null"
        businessType:
          type: string
        type:
          type: string
          enum: [category, service_shortcut, subcategory]
        bannerImageUrl:
          type:
            - string
            - "null"
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CategoryWithSubs:
      type: object
      required:
        - id
        - name
        - slug
        - icon
        - accentColor
        - businessType
        - isActive
        - sortOrder
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        accentColor:
          type: string
        parentId:
          type:
            - integer
            - "null"
        businessType:
          type: string
        type:
          type: string
          enum: [category, service_shortcut, subcategory]
        bannerImageUrl:
          type:
            - string
            - "null"
        isActive:
          type: boolean
        sortOrder:
          type: integer
        subCategories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Ad:
      type: object
      required:
        - id
        - type
        - title
        - bgColor
        - icon
        - isActive
        - sortOrder
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        subtitle:
          type:
            - string
            - "null"
        badge:
          type:
            - string
            - "null"
        bgColor:
          type: string
        accentColor:
          type:
            - string
            - "null"
        icon:
          type: string
        imageUrl:
          type:
            - string
            - "null"
        linkUrl:
          type:
            - string
            - "null"
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Short:
      type: object
      required:
        - id
        - title
        - isActive
        - sortOrder
      properties:
        id:
          type: integer
        title:
          type: string
        imageUrl:
          type:
            - string
            - "null"
        videoUrl:
          type:
            - string
            - "null"
        restaurantId:
          type:
            - integer
            - "null"
        restaurantName:
          type:
            - string
            - "null"
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ShortBody:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        imageUrl:
          type: string
        videoUrl:
          type: string
        restaurantId:
          type: integer
        restaurantName:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned by /auth/login, /auth/verify-otp, or /backend/login. Some HTML document endpoints also accept the same token as a token query parameter.